requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
argon2-cffi==20.1.0
    # via django
asgiref==3.3.4
    # via django
cffi==1.14.5
    # via argon2-cffi
dj-database-url==0.5.0
    # via -r requirements.in
django-braces==1.14.0
    # via -r requirements.in
django-model-utils==4.1.1
    # via -r requirements.in
django-widget-tweaks==1.4.8
    # via -r requirements.in
django[argon2]==3.2
    # via
    #   -r requirements.in
    #   django-braces
    #   django-model-utils
gunicorn==20.1.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pycparser==2.20
    # via cffi
python-slugify==4.0.1
    # via -r requirements.in
pytz==2021.1
    # via django
six==1.15.0
    # via
    #   argon2-cffi
    #   django-braces
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via python-slugify
wheel==0.36.2
    # via -r requirements.in
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
